app/requirements.txt
streamlit
python-docx
streamlit-aggrid
mammoth
pdfkit 
wkhtmltopdf